+++ libs/dot-components/src/lib/components/navigation/Navigation.tsx
@@ -22,9 +22,9 @@ const StyledNavigation = styled.nav`
         &.horizontal {
           flex-direction: row;
 
-          li a {
+          /* li a {
             padding: 0 ${theme.spacing(2)}px;
-          }
+          } */
         }
 
         &.vertical {